change weights to be float. Add virtual stateFilter and an implementation, goboFilter. Pass all of the states through this filter before choosing the next state.
git-svn-id: https://qlc.svn.sourceforge.net/svnroot/qlc@1245 0db28fee-5c31-0410-96ea-e69cce70e049

+++ branches/soundToLight/main/blackoutState.cpp
@@ -30,6 +30,6 @@ void BlackoutState::stop() {
 	_app->outputMap()->setBlackout(false);
 }
 
-int BlackoutState::baseWeight() {
-	return 10;
+float BlackoutState::baseWeight() {
+	return 10.0;
 }
