+++ branches/soundToLight/main/sceneState.cpp
@@ -34,6 +34,6 @@ void SceneState::stop() {
 	function->stop(_app->masterTimer());
 }
 
-int SceneState::baseWeight() {
-	return 1;
+float SceneState::baseWeight() {
+	return 1.0;
 }
